Conflicts:
	alm/src/main/java/com/dili/alm/domain/dto/NextWeeklyDto.java
	alm/src/main/java/com/dili/alm/service/WeeklyService.java
	alm/src/main/java/com/dili/alm/service/impl/WeeklyServiceImpl.java
	alm/src/main/java/com/dili/alm/utils/WordExport.java
	alm/src/main/resources/mapper/WeeklyMapper.xml
	alm/src/main/resources/templates/weekly/addWeeklyDesc.html
	alm/src/main/resources/templates/weekly/indexDesc.html

+++ alm/src/main/java/com/dili/alm/controller/FilesController.java
@@ -116,8 +116,11 @@ public class FilesController {
 			@ApiImplicitParam(name = "id", paramType = "form", value = "Files的主键", required = true, dataType = "long") })
 	@RequestMapping(value = "/delete", method = { RequestMethod.GET, RequestMethod.POST })
 	public @ResponseBody BaseOutput delete(Long id) {
-		filesService.delete(id);
-		return BaseOutput.success("删除成功");
+		Files files = this.filesService.get(id);
+		if(files!=null){
+			filesService.delete(id);
+		}
+		return BaseOutput.success("删除成功").setData(files.getName());
 	}
 
 	/**
